pyproject.toml
[project]
name = "fastapi-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aerich>=0.9.2",
    "aiohttp>=3.13.3",
    "aiomysql>=0.3.2",
    "apscheduler>=3.11.2",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "dotenv>=0.9.9",
    "email-validator>=2.3.0",
    "fastapi>=0.121.2",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "opencv-python>=4.13.0.90",
    "pandas>=2.3.3",
    "passlib>=1.7.4",
    "pycryptodome>=3.23.0",
    "pydantic>=2.12.5",
    "pyexecjs>=1.5.1",
    "python-calamine>=0.6.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=7.1.0",
    "requests>=2.32.5",
    "tortoise-orm>=0.25.3",
    "uvicorn>=0.38.0",
    "yagmail>=0.15.293",
]
[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[tool.aerich]
tortoise_orm = "main.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
